GeoLocation refactoring: android and native parts.

+++ platform/android/Rhodes/src/com/rhomobile/rhodes/RhodesActivity.java
@@ -228,7 +228,7 @@ public class RhodesActivity extends BaseActivity {
 		mHandler.post(r);
 	}
 	
-	public void post(Runnable r, int delay) {
+	public void post(Runnable r, long delay) {
 		mHandler.postDelayed(r, delay);
 	}
